backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9
apscheduler==3.10.4
pydantic[email]==2.7.1
redis==5.0.4
slowapi==0.1.9
scikit-learn==1.4.2
numpy==1.26.4
python-telegram-bot==21.3
psycopg2-binary==2.9.9
